initial react-router stuff (dont merge yet)
