+++ package.json
@@ -14,7 +14,8 @@
     "dayjs": "^1.11.4",
     "graphql": "^16.5.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "6"
   },
   "devDependencies": {
     "@tailwindcss/line-clamp": "^0.4.0",
